Move CodeBlock into converters.py
- Moves the `CodeBlock` class out of the prog.py cog into converters.py for planned reuse in other cog
